speech-to-text/requirements.txt
boto3==1.28.70
ffmpeg-python==0.2.0
fastapi==0.100.1
moviepy==1.0.3
pydantic==1.10.9
python-dotenv==1.0.0
uvicorn==0.22.0
deepgram-sdk==2.0.0
Hypercorn==0.17.3
